adding shortcuts
adding shortcuts for random, reverse and updating the display to have a to be implemented display when a shortcut has not been programmed yet. Also starting to fix multiselect.
also, fixing LEDs so that the play LED acts correctly when in CV2 edit mode. also some more work on getting CV2 to play correctly and on the correct step. man i gotta check in more frequently.

Signed-off-by: Tenkai Kariya <tenkai@zetaohm.com>