Beta 17q compilation
warning - this could be a broken build… did not commit early enough, and will be comitting half the code now, which i think was all the code that was written before I made a bunch of stupid changes, and then ill be comitting the second half after, which… doesnt really work all too well yet. maybe it will by the time I commit?

Signed-off-by: Tenkai Kariya <tenkai@zetaohm.com>